requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
python-dotenv==1.0.1
python-telegram-bot==21.4
Werkzeug==3.0.4